+++ basic/src/main/java/com/ex01/basic/service/MemberService.java
@@ -1,10 +1,9 @@
 package com.ex01.basic.service;
 
-import com.ex01.basic.dto.LoginDto;
 import com.ex01.basic.dto.MemberDto;
 import com.ex01.basic.dto.MemberRegDto;
 import com.ex01.basic.entity.MemberEntity;
-import com.ex01.basic.exception.InvalidLoginException;
+import com.ex01.basic.exception.MemberAccessDeniedException;
 import com.ex01.basic.exception.MemberDuplicationException;
 import com.ex01.basic.exception.MemberNotFoundException;
 import com.ex01.basic.repository.MemberRepository;
@@ -14,6 +13,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -24,6 +24,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class MemberService {
     private final MemberRepository memberRepository;
     private final MemberFileService memberFileService;
+    private final PasswordEncoder passwordEncoder;
 
     @Transactional(readOnly = true)
     public Page<MemberEntity> getList(int start) {
@@ -40,33 +41,57 @@ public class MemberService {
     }
 
     @Transactional(readOnly = true)
-    public MemberDto getOne(int id) {
+    public MemberDto getOne(int id, String username) {
+        MemberEntity memberEntity = memberRepository.findById(id)
+                .orElseThrow(MemberNotFoundException::new);
+        if(!memberEntity.getUsername().equals(username)){
+            throw new MemberAccessDeniedException();
+        }
         return memberRepository
                 .findById(id)
                 .map(MemberDto::new)
                 .orElseThrow(MemberNotFoundException::new);
     }
 
-    public void update(int id, MemberRegDto memberRegDto, MultipartFile multipartFile) {
+    public void update(int id,
+                       MemberRegDto memberRegDto,
+                       MultipartFile multipartFile,
+                       String username) {
         if (!memberRepository.existsById(id)) throw new MemberNotFoundException();
+
+        MemberEntity memberEntity = memberRepository.findById(id)
+                .orElseThrow(MemberNotFoundException::new);
+        if(!memberEntity.getUsername().equals(username)){
+            throw new MemberAccessDeniedException();
+        }
+
         String changeFileName = memberFileService.saveFile(multipartFile);
         if (!changeFileName.equals("nan")) {
             memberFileService.deleteFile(memberRegDto.getFileName());
             memberRegDto.setFileName(changeFileName);
         }
 
-        MemberEntity memberEntity = new MemberEntity();
         MemberDto memberDto = new MemberDto();
         BeanUtils.copyProperties(memberRegDto, memberDto);
         memberDto.setId(id);
+
+        if(!memberDto.getPassword().equals(memberEntity.getPassword()))
+            memberDto.setPassword(passwordEncoder.encode(memberRegDto.getPassword()));
         BeanUtils.copyProperties(memberDto, memberEntity);
         memberRepository.save(memberEntity);
 
     }
 
-    public void delMember(int id) {
-        if (!memberRepository.existsById(id)) throw new MemberNotFoundException();
+    public void delete(int id,
+                       String fileName,
+                       String username) {
+        MemberEntity memberEntity = memberRepository.findById(id)
+                .orElseThrow(MemberNotFoundException::new);
+        if(!memberEntity.getUsername().equals(username)){
+            throw new MemberAccessDeniedException();
+        }
         memberRepository.deleteById(id);
+        memberFileService.deleteFile(fileName);
     }
 
     public void insert(MemberRegDto memberRegDto, MultipartFile multipartFile) {
@@ -75,6 +100,7 @@ public class MemberService {
 
         String fileName = memberFileService.saveFile(multipartFile);
         memberRegDto.setFileName(fileName);
+        memberRegDto.setPassword(passwordEncoder.encode(memberRegDto.getPassword()));
         MemberEntity memberEntity = new MemberEntity();
         BeanUtils.copyProperties(memberRegDto, memberEntity);
         memberRepository.save(memberEntity);
@@ -87,11 +113,6 @@ public class MemberService {
 //        }
     }
 
-    public void login(LoginDto loginDto) {
-        boolean chck = memberRepository
-                .findByLoginInfo(loginDto.getUsername(), loginDto.getPassword())
-                .isEmpty();
-        if (chck) throw new InvalidLoginException();
-    }
+
 
 }
